requirements.txt
Flask==2.3.3
paddlepaddle==2.6.1
paddleocr==3.2.0
Pillow==10.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Werkzeug==2.3.7